198 numpy update

+++ sksurgeryvtk/models/vtk_point_model.py
@@ -52,7 +52,7 @@ class VTKPointModel(vbm.VTKBaseModel):
             raise ValueError('colours should have > 0 rows.')
         if colours.shape != points.shape:
             raise ValueError('points and colours should have the same shape.')
-        if points.dtype != np.float:
+        if points.dtype != float:
             raise TypeError('points should be float type.')
         if colours.dtype != np.byte:
             raise TypeError('colours should be byte (unsigned char) type.')
